# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Prompt Monitor Dashboard API
  description: API for monitoring AI prompts across multiple devices, providing real-time insights and compliance tracking.
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://prompt-monitor-dashboard.vercel.app
    description: Production server
...
paths:
...
  /api/prompt:
    post:
      tags:
        - Prompts
      summary: Capture a prompt
      description: Capture a single prompt from the extension. Requires authentication.
      operationId: capturePrompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CapturePromptRequest"
      responses:
        "200":
          description: Prompt captured successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CapturePromptResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/extension/status:
    post:
      tags:
        - Extension
      summary: Report extension status
      description: Report extension health status (enabled/disabled). Requires authentication.
      operationId: reportExtensionStatus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExtensionStatusRequest"
      responses:
        "200":
          description: Status reported successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtensionStatusResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required:
        - error
...
    CapturePromptRequest:
      type: object
      description: Request body for capturing a prompt
      properties:
        id:
          type: string
          description: UUID v4 message ID
        site:
          type: string
          description: Site where prompt was captured
        url:
          type: string
          description: URL where prompt was captured
        prompt:
          type: string
          description: The prompt text
        timestamp:
          type: integer
          format: int64
          description: Timestamp in epoch milliseconds
        selector:
          type: string
          nullable: true
          description: DOM selector for the prompt input
        conversation_id:
          type: string
          nullable: true
          description: Conversation ID if applicable
        trigger:
          type: string
          nullable: true
          description: What triggered the prompt capture
        page_title:
          type: string
          nullable: true
          description: Page title
        browser_name:
          type: string
          nullable: true
          description: Browser name
        browser_version:
          type: string
          nullable: true
          description: Browser version
        profile_name:
          type: string
          nullable: true
          description: Browser profile name
        username:
          type: string
          nullable: true
          description: Username
        agent_status:
          type: string
          nullable: true
          description: Status of native agent
        instance_id:
          type: string
          nullable: true
          description: Browser instance ID
        extension_version:
          type: string
          nullable: true
          description: Extension version
      required:
        - id
        - site
        - url
        - prompt
        - timestamp
    CapturePromptResponse:
      type: object
      properties:
        status:
          type: string
          description: Capture status
        prompt_id:
          type: string
          description: Captured prompt ID
        message:
          type: string
          description: Success message
      required:
        - status
        - prompt_id
        - message
...
    ExtensionStatusRequest:
      type: object
      description: Request body for extension status report
      properties:
        instance_id:
          type: string
          description: Browser instance ID
        status:
          type: string
          description: Extension status (e.g., enabled, disabled)
        reported_by:
          type: string
          nullable: true
          description: Who/what reported the status
        metadata:
          type: object
          nullable: true
          description: Additional metadata
      required:
        - instance_id
        - status
    ExtensionStatusResponse:
      type: object
      properties:
        status:
          type: string
          description: Report status
        status_id:
          type: string
          description: Created status record ID
        message:
          type: string
          description: Success message
      required:
        - status
        - status_id
        - message
...