# src/generated/api/models/CapturePromptResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CapturePromptResponse = {
  /**
   * Capture status
   */
  status: string;
  /**
   * Captured prompt ID
   */
  prompt_id: string;
  /**
   * Success message
   */
  message: string;
};