# src/generated/api/models/ExtensionStatusResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ExtensionStatusResponse = {
  /**
   * Report status
   */
  status: string;
  /**
   * Created status record ID
   */
  status_id: string;
  /**
   * Success message
   */
  message: string;
};